% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allunique.R
\name{allunique}
\alias{allunique}
\title{allunique}
\usage{
allunique(x)
}
\arguments{
\item{x}{is a vector}
}
\value{
a logical value
}
\description{
Determines whether a vector contains only unique, non-null values
}
\examples{
allunique(c("A","A","B"))

}
