% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATsqlQuery.R
\name{BATsqlQuery}
\alias{BATsqlQuery}
\title{BATsqlQuery}
\usage{
BATsqlQuery(
  ids,
  table = "PLOT_OWNER",
  field = "CN",
  idtag = "<I>",
  ttag = "<T>",
  ftag = "<F>",
  query = "SELECT * FROM FS_NWOS.<T> WHERE <F> IN ('<I>')",
  conn = "con"
)
}
\arguments{
\item{ids}{is a character vector containing NWOS CNs to be queried.}

\item{table}{is a character string containing the name of the table to be queried.}

\item{field}{is a character string containing the name of the field in the desired table containing the desired CNs.}

\item{idtag}{is a character string containing the name of the tag in the query template that refers to the ids (for use with custom queries).}

\item{ttag}{is a character string containing the name of the tag in the query template that refers to the table (for use with custom queries).}

\item{ftag}{is a character string containing the name of the tag in the query template that refers to the field (for use with custom queries).}

\item{query}{is a character string containing text of a custom query, containing the values of idtag, ttag, and ftag.}

\item{conn}{is a character string containing the name of the open RODBC channel.}
}
\description{
Queries the NWOS DB to select batches of specific records from tables or views
}
\details{
This function must be run inside an open RODBC channel (32-bit only) connected to the USFS FIA production database through a user with write permissions.
}
\examples{
BATsqlQuery(ids=c('SAM0001','SAM0002',table=c('SAMPLE'))

}
