% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BATsqlQuery64.R
\name{BATsqlQuery64}
\alias{BATsqlQuery64}
\title{BATsqlQuery64}
\usage{
BATsqlQuery64(ids, table = "PLOT_OWNER", field = "CN", idtag = "<I>",
  ttag = "<T>", ftag = "<F>",
  query = "SELECT * FROM FS_NWOS.<T> WHERE <F> IN ('<I>')",
  conn = "con")
}
\arguments{
\item{ids}{is a character vector containing NWOS CNs to be queried.}

\item{table}{is a character string containing the name of the table to be queried.}

\item{field}{is a character string containing the name of the field in the desired table containing the desired CNs.}

\item{idtag}{is a character string containing the name of the tag in the query template that refers to the ids (for use with custom queries).}

\item{ttag}{is a character string containing the name of the tag in the query template that refers to the table (for use with custom queries).}

\item{ftag}{is a character string containing the name of the tag in the query template that refers to the field (for use with custom queries).}

\item{query}{is a character string containing text of a custom SQL query, containing the values of idtag, ttag, and ftag.}
}
\value{
a dataframe containing the DB return
}
\description{
Wrapper function for RODBC:sqlQuery to select batches of specific records from tables or views
}
\details{
This function must be run on a machine with an ODBC connection (32-bit) to the USFS FIA production database through a user with read permissions.
}
\examples{
BATsqlQuery64(ids=c('SAM0001','SAM0002',table=c('SAMPLE'))

}
