% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cg_max.R
\name{cg_max}
\alias{cg_max}
\title{cg_max}
\usage{
cg_max(
  check_group = "ACQ_TYPE",
  quest_wide = QUEST_WIDE,
  quest_meta = QUEST_META
)
}
\description{
Calculate maximum values for check group questions
}
\details{
For area and cooperation rate tables see ...
}
\examples{
cg_max("ACQ_TYPE")

}
