% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edt.R
\name{edt}
\alias{edt}
\title{edt}
\usage{
edt(x)
}
\arguments{
\item{x}{is a character vector of in MS Excel DATETIME format}
}
\value{
a date vector
}
\description{
Converts MS Excel DATETIME to Date
}
\examples{
edt(43383)

}
