% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_table.R
\name{expand_table}
\alias{expand_table}
\title{expand_table}
\usage{
expand_table(x, col)
}
\arguments{
\item{x}{is a nwos.object}

\item{col}{is string referencing the numeric column containing record-level expansion factors}
}
\value{
a data.frame
}
\description{
A function for replicating each record in a dataframe by an expansion factor stored as a separate column in that dataframe
}
\examples{
expand_table(foo,col='NUM_PLOTS')

}
