% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_update_table.r
\name{export_update_table}
\alias{export_update_table}
\title{export_update_table}
\usage{
export_update_table(df)
}
\arguments{
\item{df}{is a string containing the name of a NWOS-DB update table}
}
\description{
Exports an update table dataframe as a CSV along with a SQL file to the T drive, to manually update NWOS-DB
}
\details{
This function must be run on a machine with a connection to USFS T drive
}
\examples{
export_update_table(df="SAMPLE_UPDATE")

}
