% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{get_mode}
\usage{
get_mode(x)
}
\arguments{
\item{x}{is a vector}
}
\description{
Find the modal value for a vector
}
\details{
Finds the most common value. If there is a tie, the value that
is first alphabetically is taken.
}
\examples{
get_mode(c("a", "a", "b"))
get_mode(c("a", "b", "b"))
get_mode(c("a", "a", "b", "b"))
get_mode(c("b", "b", "a", "a"))

}
