% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwos.R
\name{get_nwos}
\alias{get_nwos}
\title{get_nwos}
\usage{
get_nwos(
  cycle = "2018",
  study = "base",
  states = NA,
  questions = NA,
  strict.intensification = FALSE
)
}
\arguments{
\item{cycle}{is a string containing the NWOS cycle desired.}

\item{study}{is a string containing the NWOS study desired.}

\item{states}{is a character vector containing one or more states to extract.}

\item{questions}{is a character vector containing the names of one or more NWOS questions to extract.}

\item{strict.intensification}{is a logical value determining whether ONLY intensified plots should be selected.}
}
\value{
a nwos.object
}
\description{
Extracts the data for a given NWOS cycle/study from the NWOS and returns it as an nwos.object, including metadata, weights, imputations, and plot-level data.
}
\details{
This function must be run on a machine with an ODBC connection to the USFS FIA production database through a user with read permissions.
}
\examples{
get_nwos(cycle='2018',study='base',states='44',questions='AC_LAND',strict.intensification=FALSE)

}
