% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwos_ancillary.r
\name{get_nwos_ancillary}
\alias{get_nwos_ancillary}
\title{get_nwos_ancillary}
\usage{
get_nwos_ancillary(var = NA, by = "all")
}
\arguments{
\item{var}{is a string containing the name of the ancillary variable to be returned.}

\item{by}{is a function for aggregating point-level values, e.g. 'mean','min','max', etc.}
}
\value{
a data.frame
}
\description{
Returns ancillary data, in raw or aggregated form, for use with raw NWOS data.
}
\details{
This function must be run on a machine with an ODBC connection to the USFS FIA production database through a user with read permissions. If the function is run with no variable listed in the ‘var’ parameter, it will return the metadata for all ancillary datasets in the database. This table is useful as a lookup table. Otherwise, the ‘by’ parameter determines how the returned data are aggregated. If we select ‘by=all’ (the default), the raw values are returned for all sample points. For numeric variables, however, we will usually want to aggregate the values by survey (for surveys with more than one sample point) in order to attach to a wide NWOS table.
}
\examples{
get_nwos_ancillary(var="CENSUS",by="all")
get_nwos_ancillary(var="ACREAGE",by="mean")

}
