% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwos_db_catalog.R
\name{get_nwos_db_catalog}
\alias{get_nwos_db_catalog}
\title{get_nwos_db_catalog}
\usage{
get_nwos_db_catalog(dir = ".", file = "NWOS_DB_catalog")
}
\arguments{
\item{dir}{is a string vector containing the path where the catalog will be generated.}

\item{file}{is a string vector containing the file name (without file extension) that will be generated.}
}
\description{
Generates NWOS database catalog
}
\details{
This function generates an up-to-date HTML version of the NWOS database catalog.
}
\examples{
get_nwos_db_catalog(dir='C:/foo',file='NWOS_DB_catalog')

}
