% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwos_plots.R
\name{get_nwos_plots}
\alias{get_nwos_plots}
\title{get_nwos_plots}
\usage{
get_nwos_plots(cycle = "2018", study = "base")
}
\arguments{
\item{cycle}{is a string containing the NWOS cycle desired.}

\item{study}{is a string containing the NWOS study desired.}
}
\value{
an nwos.plots object
}
\description{
Extracts the full plot list for a given NWOS cycle/study, classed to STRATA
}
\details{
This function must be run on a machine with an ODBC connection to the USFS FIA production database through a user with read permissions.
}
\examples{
get_nwos_plots(cycle='2018',study='base')

}
