% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwos_plots_islands.R
\name{get_nwos_plots_islands}
\alias{get_nwos_plots_islands}
\title{get_nwos_plots_islands}
\usage{
get_nwos_plots_islands(cycle = "2023", study = "islands", geo = NA,
  yrs = NA)
}
\arguments{
\item{cycle}{is a string containing the NWOS cycle desired.}

\item{study}{is a string containing the NWOS study desired.}

\item{geo}{is a string containing the NWOS statecds desired.}

\item{yrs}{is a vector containing the specific years desired with the NWOS cycle.}
}
\value{
an nwos.plots object
}
\description{
Extracts the full plot list for a given NWOS cycle/study, classed to STRATA
}
\details{
This function must be run on a machine with an ODBC connection to the USFS FIA production database through a user with read permissions.
}
\examples{
get_nwos_plots_islands(cycle='2023',study='islands',geo='78',yrs='2019')

}
