% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwos_w_incomplete.R
\name{get_nwos_w_incomplete}
\alias{get_nwos_w_incomplete}
\title{get_nwos}
\usage{
get_nwos_w_incomplete(cycle = "2018", study = "base", states = NA,
  questions = NA)
}
\arguments{
\item{cycle}{is a string containing the NWOS cycle desired.}

\item{study}{is a string containing the NWOS study desired.}

\item{states}{is a character vector containing one or more states to extract.}

\item{questions}{is a character vector containing the names of one or more NWOS questions to extract.}
}
\value{
a nwos.object
}
\description{
Extracts the data for a given NWOS cycle/study from the NWOS and returns it as an nwos.object, including metadata, weights, and imputations.
}
\details{
This function must be run on a machine with an ODBC connection to the USFS FIA production database through a user with read permissions.
}
\examples{
get_nwos(cycle='2018',study='base',states='44',questions='AC_LAND')

}
