% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwos_weights.R
\name{get_nwos_weights}
\alias{get_nwos_weights}
\title{get_nwos_weights}
\usage{
get_nwos_weights(cycle = "2018", strata = "FFO")
}
\arguments{
\item{cycle}{is a string containing the NWOS cycle desired.}

\item{strata}{is a string containing the name of the strata desired.}
}
\description{
Downloads replicate weights for NWOS samples
}
\details{
This function must be run on a computer connected to the USFS T drive, with appropriate read permissions
}
\examples{
get_nwos_weights(cycle='2018',strata='FFO')

}
