% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunique.R
\name{lunique}
\alias{lunique}
\title{lunique}
\usage{
lunique(x)
}
\arguments{
\item{x}{is a vector}
}
\value{
a numeric value
}
\description{
Calculates number of unique values in a vector
}
\examples{
lunique(c("A","A","B"))

}
