% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcn.R
\name{maxcn}
\alias{maxcn}
\title{maxcn}
\usage{
maxcn(x)
}
\arguments{
\item{x}{is a character vector of NWOS CNs}
}
\value{
a numeric vector
}
\description{
Identifies NWOS CN with the highest numerical component
}
\examples{
maxcn(c('SAM00001','SAM04567'))

}
