% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{mode}
\usage{
mode(x)
}
\arguments{
\item{x}{is a vector}
}
\description{
Find the modal value for a vector
}
\details{
Finds the most common value. If there is a tie, the value that
is first alphabetically is taken.
}
\examples{
mode(c("a", "a", "b"))
mode(c("a", "b", "b"))
mode(c("a", "a", "b", "b"))
mode(c("b", "b", "a", "a"))

}
