% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncn.R
\name{ncn}
\alias{ncn}
\title{ncn}
\usage{
ncn(x)
}
\arguments{
\item{x}{is a character vector of NWOS CNs}
}
\value{
a numeric vector
}
\description{
Extracts numeric component of NWOS CN
}
\examples{
ncn(c('SAM00001','OWN04567'))

}
