% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullif.R
\name{nullif}
\alias{nullif}
\title{nullif}
\usage{
nullif(x)
}
\arguments{
\item{x}{is a vector with null and nonnull values}
}
\value{
a character vector with null values replaced by an empty string
}
\description{
Set null values to empty strings
}
\examples{
nullif(c(1,2,NA))

}
