% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos.object.R
\docType{class}
\name{nwos.object-class}
\alias{nwos.object-class}
\alias{nwos.object}
\title{nwos.object}
\description{
An S4 class to contain NWOS raw data, including metadata,weights,and imputations.
}
\section{Slots}{

\describe{
\item{\code{quest}}{is a data.frame containing raw questionaire data}

\item{\code{sample}}{is a data.frame containing imformation on the sample}

\item{\code{metadata}}{is a data.frame containing question metadata}

\item{\code{fields}}{is a data.frame containing metadata for columns in sample and quest}

\item{\code{weights}}{is a data.frame containing survey weights}

\item{\code{imputations}}{is a data.frame containing imputation sets}
}}

