% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos.plots.object.R
\docType{class}
\name{nwos.plots.object-class}
\alias{nwos.plots.object-class}
\alias{nwos.plots.object}
\title{nwos.plots.object}
\description{
An S4 class to contain NWOS raw plot- and sample-level data
}
\section{Slots}{

\describe{
\item{\code{plots}}{is a data.frame containing raw plot-level data for a cycle}

\item{\code{response_codes}}{is a data.frame containing response codes and other sample-level data to attach to plots}
}}

