% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_cooperation_rate.R
\name{nwos_cooperation_rate}
\alias{nwos_cooperation_rate}
\title{nwos_cooperation_rate}
\usage{
nwos_cooperation_rate(x)
}
\arguments{
\item{x}{is a vector containing values of I ('response'), P ('partial response'), NC ('not contacted'), UN ('unknown'), and R ('refused')}
}
\value{
a numeric value
}
\description{
Calculates nwos cooperation rate (~AAPOR's COOP1)
}
\examples{
nwos_cooperation_rate(c('I','I','R','R','NC','NC','P'))

}
