% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_data_recode.R
\name{nwos_data_recode}
\alias{nwos_data_recode}
\title{NWOS Data Recode}
\usage{
nwos_data_recode(data = QUEST_WIDE, meta.data = QUEST_META, box2 = c(names(data)[grepl("OBJ_", names(data))], names(data)[grepl("CNC_", names(data))], "ATT_WOODED", "ATT_SELL", "KNOW_WOOD", "WANT_KNOW_WOOD", "EMO_WOOD", "WOOD_COMMUNITY"))
}
\arguments{
\item{data}{= QUEST_WIDE,}

\item{meta.data}{= QUEST_META,}

\item{box2}{= c(names(data)[grepl("OBJ_", names(data))], names(data)[grepl("CNC_", names(data))], "ATT_WOODED", "ATT_SELL", "KNOW_WOOD", "WANT_KNOW_WOOD", "EMO_WOOD", "WOOD_COMMUNITY")}
}
\description{
Recode and add basic NWOS variables.
}
\details{
The default values create the variables used in the NWOS tables.
}
\examples{
??
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
