% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates.R
\name{nwos_estimates}
\alias{nwos_estimates}
\title{NWOS Estimates Total}
\usage{
nwos_total_svl(state, variable, level, data = QUEST, area = NA, weights = "WEIGHT")
}
\arguments{
\item{data}{= QUEST_WIDE}

\item{area}{= NA}

\item{weights}{= "WEIGHT"}

\item{state}{= "1"}

\item{variable}{= "HOME"}

\item{level}{= "1",}
}
\description{
Estimator used to calculate NWOS totals by state, variable and level.
}
\examples{
state = "1"
variable = "OWNTYPE"
level = "1"
data = QUEST[[1]]
area = NA
weights = "WEIGHT"
stratum = NA
domain = NA
??
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
