% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_add_minority.R
\name{nwos_estimates_add_minority}
\alias{nwos_estimates_add_minority}
\title{Add MINORITY Variable to NWOS Dataset}
\usage{
nwos_estimates_data_add_variables(data = QUEST_WIDE, meta.data = QUEST_META, box2 = c(names(data)[grepl("OBJ_", names(data))], names(data)[grepl("CNC_", names(data))], "ATT_WOODED", "ATT_SELL", "KNOW_WOOD", "WANT_KNOW_WOOD", "EMO_WOOD", "WOOD_COMMUNITY"))
}
\arguments{
\item{data}{= QUEST_WIDE}

\item{minority}{Logical value indicating if the OWN1_MINORITY variable should be created. Default = F}

\item{tele}{Logical value indicating if the TELE related variables should be created. Default = F}
}
\description{
Add variables to an NWOS dataframe
}
\details{
The default values create the variables used in the NWOS tables.
}
\examples{
nwos_estimates_data_add_variables(minority = T, tele = T)
}
\keyword{nwos}
