% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_add_tele.R
\name{nwos_estimates_add_tele}
\alias{nwos_estimates_add_tele}
\title{Add TELE Variables to NWOS Dataset}
\usage{
nwos_estimates_data_add_variables(data = QUEST_WIDE, meta.data = QUEST_META, box2 = c(names(data)[grepl("OBJ_", names(data))], names(data)[grepl("CNC_", names(data))], "ATT_WOODED", "ATT_SELL", "KNOW_WOOD", "WANT_KNOW_WOOD", "EMO_WOOD", "WOOD_COMMUNITY"))
}
\arguments{
\item{x}{list number. Only applicable if is data is a list of data frames, instead of a single data frame. This used mainly for apply functions.}

\item{data}{data frame or list of data frames}
}
\description{
Add variables to an NWOS dataframe
}
\details{
The default values create the variables used in the NWOS tables.
}
\examples{
nwos_estimates_data_add_variables(minority = T, tele = T)
}
\keyword{nwos}
