% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_add_variable.R
\name{nwos_estimates_add_variable}
\alias{nwos_estimates_add_variable}
\title{Add Variables to NWOS QUEST objects}
\usage{
nwos_estimates_add_variable(x = 1, data = QUEST_LIST, variable.name = "TEN_PLUS", variable.definition = "AC_WOOD >= 10")
}
\arguments{
\item{x}{list number. Only applicable if is data is a list of data frames, instead of a single data frame. This used mainly for apply functions.}

\item{data}{list of NWOS QUEST objects}
}
\description{
Add variables to an NWOS data frame
}
\details{
Add a new variable to all NWOS QUEST objects in a list.
}
\examples{
nwos_estimates_add_variable()
}
\keyword{nwos}
