% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_categorical.R
\name{nwos_estimates_categorical}
\alias{nwos_estimates_categorical}
\title{nwos_estimates_categorical}
\usage{
nwos_estimates_categorical(
  variable.name = "MAN_PLAN",
  level.name = "1",
  domain.name = "AC_WOOD >= 10"
)
}
\arguments{
\item{variable.name}{= "MAN_PLAN",}

\item{level.name}{= "1",}

\item{domain.name}{= "AC_WOOD >= 10"}
}
\value{
data frame
}
\description{
Generate estimates for categorical variables
This is functuion is designed to be called by nwos_estimates_geo
}
\examples{
nwos_estimates_categorical()

}
