% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_combine_geo.R
\name{nwos_estimates_combine_geo}
\alias{nwos_estimates_combine_geo}
\title{nwos_estimates_summary}
\usage{
nwos_estimates_combine_geo(domain = "FFO", stratum = "TENPLUS")
}
\description{
Combine and ummarize Imputations
}
\examples{
# Check for missings reps
files <- list.files("DATA/FFO/TENPLUS/REP_GEO", pattern = "_US.RDS")
rep.nums <- 0:1000
x <- gsub("FFO_TENPLUS_", "", gsub("_US.RDS", "", files))
rep.nums[!rep.nums \%in\% x]
domain = "FFO"
stratum = "ONEPLUS"
nwos_estimates_summary()

}
