% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_continuous.R
\name{nwos_estimates_continuous}
\alias{nwos_estimates_continuous}
\title{nwos_estimates_continuous}
\usage{
nwos_estimates_continuous(
  variable.name = "AC_WOOD",
  domain.name = "AC_WOOD >= 10"
)
}
\arguments{
\item{variable.name}{= "AC_WOOD",}

\item{domain.name}{= "AC_WOOD >= 10"}
}
\value{
data frame
}
\description{
Generate estimates for continuous variables
This is functuion is designed to be called by nwos_estimates_geo
}
\examples{
nwos_estimates_continuous()

}
