% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_geo.R
\name{nwos_estimates_geo}
\alias{nwos_estimates_geo}
\title{nwos_estimates_geo}
\usage{
nwos_estimates_geo(
  domain = "AC_WOOD >= 10",
  geo = GEO,
  cond.status = 1,
  own = 45,
  quest = QUEST,
  quest.meta = QUEST_META,
  imputations = 1:5,
  variables.categorcial = VARIABLES_CATEGORICAL,
  variables.continuous = VARIABLES_CONTINUOUS
)
}
\arguments{
\item{domain}{= "AC_WOOD >= 10",}

\item{geo}{= GEO,}

\item{cond.status}{= 1,}

\item{own}{= 45,}

\item{quest}{= QUEST,}

\item{imputations}{= 1:5}

\item{plot}{= PLOTS,}
}
\value{
data frame
}
\description{
Generate estimates by geography
}
\examples{
nwos_estimates_geo(imputations = 1:2)
nwos_estimates_geo(imputations = 1:2, domain = "AC_WOOD >= 1")
nwos_estimates_geo(domain = "AC_WOOD >= 1")

}
