% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_make_quest.R
\name{nwos_estimates_make_quest}
\alias{nwos_estimates_make_quest}
\title{nwos_estimates_make_quest}
\usage{
nwos_estimates_make_quest(
  imp = 1,
  quest = QUEST,
  own = 45,
  variables.categorcial = VARIABLES_CATEGORICAL,
  variables.continuous = VARIABLES_CONTINUOUS
)
}
\arguments{
\item{quest}{= QUEST}

\item{variables.categorcial}{= VARIABLES_CATEGORICAL,}

\item{variables.continuous}{= VARIABLES_CONTINUOUS)}
}
\value{
tibble
}
\description{
Convert nwos.object to wide tables - separate tables for each umputation
}
\examples{
nwos_estimates_make_quest()

}
