% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_rep_check.R
\name{nwos_estimates_rep_check}
\alias{nwos_estimates_rep_check}
\title{nwos_estimates_rep_check}
\usage{
nwos_estimates_rep_check(
  file.dir = "DATA",
  stratum = "FFO",
  domain = "ONEPLUS"
)
}
\description{
Check number of replicates
}
