% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_summary.R
\name{nwos_estimates_summary}
\alias{nwos_estimates_summary}
\title{nwos_estimates_summary}
\usage{
nwos_estimates_summary(
  geo.list = GEO_LIST,
  stratum = "FFO",
  domain = "TENPLUS"
)
}
\description{
Combine and ummarize Imputations
}
\examples{
domain = "FFO"
stratum = "ONEPLUS"
# Need to run nwos_estimates_combine first
nwos_estimates_summary(geo.list = GEO_LIST[55,])

}
