% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_summary_2.R
\name{nwos_estimates_summary_2}
\alias{nwos_estimates_summary_2}
\title{nwos_estimates_summary}
\usage{
nwos_estimates_summary_2(data)
}
\description{
Combine and ummarize Imputations
}
\examples{
domain = "FFO"
stratum = "ONEPLUS"
# Need to run nwos_estimates_combine first
nwos_estimates_summary(geo.list = GEO_LIST[55,])

}
