% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_full.R
\name{nwos_full}
\alias{nwos_full}
\title{nwos_full}
\usage{
nwos_full(quest, plots, imputations = "none")
}
\arguments{
\item{quest}{is a nwos.object}

\item{plots}{is a nwos.plots.object}

\item{imputations}{is string containing which imputation set is to be used in place of non-response, if any. Valid responses are "none","random","1","2","3","4","5"}
}
\value{
a data.frame
}
\description{
Combines nwos plots and response data to recreate the full sample
}
\examples{
nwos_full(quest,plots,imputations="1")

}
