% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_full_export.R
\name{nwos_full_export}
\alias{nwos_full_export}
\title{nwos_full_export}
\usage{
nwos_full_export(quest, plots, dir = "~", imputations = "none")
}
\arguments{
\item{quest}{is a nwos.object}

\item{plots}{is a nwos.plots.object}

\item{dir}{is the directory to export to}

\item{imputations}{is string containing which imputation set is to be used in place of non-response, if any. Valid responses are "none","random","1","2","3","4","5"}
}
\value{
a data.frame
}
\description{
Exports full NWOS dataset and metadata as CSVs, in a wide format.
}
\examples{
nwos_full_export(foo,foo2,dir='~',format='long',imputations="1")

}
