% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_full_metadata.R
\name{nwos_full_metadata}
\alias{nwos_full_metadata}
\title{nwos_full_metadata}
\usage{
nwos_full_metadata(x)
}
\arguments{
\item{x}{is a nwos.object}
}
\value{
a data.frame
}
\description{
Creates metadata table for full (i.e. all plots and responses) NWOS dataset
}
\examples{
nwos_full_metadata(foo)

}
