% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_html_table.R
\name{nwos_html_table}
\alias{nwos_html_table}
\title{nwos_html_table}
\usage{
nwos_html_table(
  tab.num,
  data = table.data,
  data.coop = coop.table.data,
  stratum.abb = STRATUM_ABB,
  stratum.name = STRATUM_NAME,
  domain.abb = DOMAIN_ABB,
  domain.name = DOMAIN_NAME,
  year = YEAR,
  inc.ac = F
)
}
\description{
nwos_html_table
}
\details{
For area and cooperation rate tables see ...
tab.num = 3
nwos_table()
nwos_table_area()
nwos_table_coop()
}
