% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_html_table_set.R
\name{nwos_html_table_set}
\alias{nwos_html_table_set}
\title{nwos_html_table_set}
\usage{
nwos_html_table_set(
  geo.abb,
  data = ESTIMATES,
  data.coop = COOP,
  ref.geo = REF_GEO,
  ref.table = REF_TABLE,
  stratum = STRATUM,
  stratum.abb = STRATUM_ABB,
  stratum.name = STRATUM_NAME,
  domain = DOMAIN,
  domain.abb = DOMAIN_ABB,
  domain.name = DOMAIN_NAME,
  year = YEAR,
  excludes = EXCLUDES,
  inc.ac = F
)
}
\description{
Create the body of an NWOS table
}
\details{
For area and cooperation rate tables see ...
}
\examples{
nwos_table_set(geo.abb = "BALT")
geo.abb = "BALT"
data = ESTIMATES
ref.geo = REF_GEO
ref.table = REF_TABLE
stratum = STRATUM
stratum.abb = STRATUM_ABB
stratum.name = STRATUM_NAME
domain = DOMAIN
domain.abb = DOMAIN_ABB
domain.name = DOMAIN_NAME
year = YEAR
excludes = EXCLUDES
inc.ac = F

}
