% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_long.R
\name{nwos_long}
\alias{nwos_long}
\title{nwos_long}
\usage{
nwos_long(x)
}
\arguments{
\item{x}{is a nwos.object}
}
\value{
a data.frame
}
\description{
Formats NWOS data into a long format for analysis
}
\examples{
nwos_long(foo)

}
