% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_long_metadata.R
\name{nwos_long_metadata}
\alias{nwos_long_metadata}
\title{nwos_long_metadata}
\usage{
nwos_long_metadata(x)
}
\arguments{
\item{x}{is a nwos.object}
}
\value{
a data.frame
}
\description{
Creates metadata table for NWOS data in long format
}
\examples{
nwos_long_metadata(foo)

}
