% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_make_replicates.R
\name{nwos_make_replicates}
\alias{nwos_make_replicates}
\title{NWOS Make Replicates}
\usage{
nwos_replicates(index, point.count, R = 2500)
}
\arguments{
\item{index}{vector of names (e.g., row names) used to identify observations.}

\item{point.count}{the number of sample points associated with each observation. Default is 1.}

\item{R}{number of replicates or bootstraps. Default is 2500.}
}
\value{
List containing indices for the observations that make up each of the replicates.
}
\description{
Generate replicates that can be used for NWOS bootstrapping variance estimation.
}
\examples{
wi <- tbl_df(read.csv("data/wi.csv")) \%>\% mutate(ROW_NAME = row.names(wi), AC_WOOD = ACRES_FOREST, FFO = if_else(LAND_USE == 1 & OWN_CD == 45 & AC_WOOD >= 1, 1, 0))
WI_REPLICATES <- nwos_replicates(index = row.names(wi), point.count = wi$POINT_COUNT, R = 100)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
