% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_mean.R
\name{nwos_mean}
\alias{nwos_mean}
\title{NWOS Mean}
\usage{
nwos_mean(weight, area = 1, domain = 1, base = 1, variable)
}
\arguments{
\item{weight}{vector of weights per ownership.}

\item{area}{vector of area (e.g., forest acres) per ownership. Default = 1 (i.e., estimates are in terms of ownerships).}

\item{domain}{vector with 1 indicating inclusion in the domain and 0 otherwise. Default = 1 (i.e., all ownerships are included).}

\item{variable}{vector of variable of interest.}
}
\description{
This function calculates means for the NWOS.
}
\examples{
nwos_mean(weight = wi$WEIGHT, variable = wi$Y_3)
nwos_mean(weight = wi$WEIGHT, area = wi$AC_WOOD, variable = wi$Y_3)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
