% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_mean_apply.R
\name{nwos_mean_apply}
\alias{nwos_mean_apply}
\title{NWOS Proportion}
\usage{
nwos_proportion_replicates(r, index.rep, data, weight.rep, owner.area.name = "OWNER", domain.name, base.name = "FFO", variable.name = "FFO")
}
\arguments{
\item{r}{vector of replicates. See details below.}

\item{index.rep}{list of observations (i.e., replicates) in data to include.}

\item{data}{data frame containing stratum variable.}

\item{weight.rep}{list with each element the weights for the observations in that replicate.}

\item{owner.area.name}{nae of a variable in data of the area (of forestland) owned by each ownership. Default = "OWNER" (i.e., 1, which is used for estimating number of ownerships).}

\item{domain.name}{name of the variable in data indicating inclusion (1) and exclusion (0) in the domain of interest.}

\item{base.name}{name of a variable in data with 1 indicating inclusion in the base (i.e., denominator) and 0 otherwise. Default = 1 (i.e., all ownerships are included).}

\item{variable.name}{name of the variable of interest in data. Default = "FFO".}
}
\description{
This function calculates proportions for the NWOS replicates. This is typically used with an apply function.
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
