% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_n.R
\name{nwos_n}
\alias{nwos_n}
\title{NWOS n}
\usage{
nwos_n(point.count = 1, stratum = 1, domain = 1)
}
\arguments{
\item{point.count}{vector of number of sample points associated with each observation. Default = 1.}

\item{stratum}{vector with 1 indicating inclusion in the stratum and 0 otherwise. Default = 1 (i.e., all ownerships are in the same stratum).}

\item{domain}{vector with 1 indicating inclusion in the domain and 0 otherwise. Default = 1 (i.e., all ownerships are included).}

\item{variable}{vector of variable of interest. Default = 1 (i.e., variable is ignored)}
}
\description{
Count number of responses (n).
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
