% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_n_svl.R
\name{nwos_n_svl}
\alias{nwos_n_svl}
\title{NWOS Sample Sizes by State, Variable, and Level}
\usage{
nwos_n_svl(states, variables, levels, data = QUEST_WIDE)
}
\arguments{
\item{data}{= QUEST_WIDE}

\item{states}{= "1"}

\item{variables}{= "HOME"}

\item{levels}{= "1",}
}
\description{
Calculate NWOS sample sizes by state, variable and level.
}
\examples{
??
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
