% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_object.R
\name{nwos_object}
\alias{nwos_object}
\title{nwos_object}
\usage{
nwos_object(x)
}
\arguments{
\item{x}{is a list}
}
\value{
a nwos.object
}
\description{
Creates an object of S4 class 'nwos.object'
}
\examples{
nwos_object(foo)

}
