% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_plot_response_rate.R
\name{nwos_plot_response_rate}
\alias{nwos_plot_response_rate}
\title{NWOS Plot Response Rate}
\usage{
nwos_plot_response_rate(stratum, point.count, response)
}
\arguments{
\item{stratum}{vector indicating inclusion (1) and exclusion (0) in the stratum of interest.}

\item{point.count}{vector of the number of sample points associated with each observation. Default is 1.}

\item{response}{vector indicating response (1) and non-response (0).}
}
\value{
Response rate in the stratum.
}
\description{
This function calculates response rates for the NWOS.
}
\examples{
wi <- tbl_df(read.csv("data/wi.csv")) \%>\% mutate(ROW_NAME = row.names(wi), AC_WOOD = ACRES_FOREST, FFO = if_else(LAND_USE == 1 & OWN_CD == 45 & AC_WOOD >= 1, 1, 0), RESPONSE = if_else(RESPONSE_PROPENSITY >= 0.5, 1, 0), RESPONSE = if_else(is.na(RESPONSE_PROPENSITY), 0, RESPONSE))
WI_FFO_RR <- nwos_response_rate(stratum = wi$FFO, point.count = wi$POINT_COUNT, response = wi$RESPONSE)
WI_FFO_RR
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
