% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_plot_response_rate_state.R
\name{nwos_plot_response_rate_state}
\alias{nwos_plot_response_rate_state}
\title{NWOS Plot Response Rate by State}
\usage{
nwos_plot_response_rate(stratum, point.count, response)
}
\arguments{
\item{state}{= "1"}

\item{data}{= po}

\item{s.name}{= "FFO"}

\item{r.name}{= "RESPONSE"}
}
\value{
Response rate in the stratum calculated by state.
}
\description{
This function calculates response rates for the NWOS by state.
}
\examples{
??
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
