% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_plots_complete.R
\name{nwos_plots_complete}
\alias{nwos_plots_complete}
\title{nwos_plots_complete}
\usage{
nwos_plots_complete(x)
}
\arguments{
\item{x}{is a nwos.plots object}
}
\value{
a data.frame
}
\description{
Creates a 'completed' form of the NWOS plots table
}
\examples{
nwos_plots_complete(foo)

}
