% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_plots_metadata.R
\name{nwos_plots_metadata}
\alias{nwos_plots_metadata}
\title{nwos_plots_metadata}
\usage{
nwos_plots_metadata(x)
}
\arguments{
\item{x}{is a nwos.plots.object}
}
\value{
a data.frame
}
\description{
Creates metadata table for NWOS plots data
}
\examples{
nwos_plots_metadata(foo)

}
