% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_plots_object.R
\name{nwos_plots_object}
\alias{nwos_plots_object}
\title{nwos_plots_object}
\usage{
nwos_plots_object(x)
}
\arguments{
\item{x}{is a list}
}
\value{
a nwos.plots.object
}
\description{
Creates an object of S4 class 'nwos.plots'
}
\examples{
nwos_plots_object(foo)

}
