% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_proportion.R
\name{nwos_proportion}
\alias{nwos_proportion}
\title{NWOS Proportion}
\usage{
nwos_proportion(weight, area = 1, domain, base = 1, variable = 1)
}
\arguments{
\item{weight}{vector of weights per ownership.}

\item{area}{vector of area (e.g., forest acres) per ownership. Default = 1 (i.e., estimates are in terms of ownerships).}

\item{domain}{vector with 1 indicating inclusion in the domain and 0 otherwise. Default = 1 (i.e., all ownerships are included).}

\item{base}{vector with 1 indicating inclusion in the base (i.e., denominator) and 0 otherwise. Default = 1 (i.e., all ownerships are included).}

\item{variable}{vector of variable of interest. Default = 1 (i.e., variable is ignored).}
}
\description{
This function calculates proportions for the NWOS.
}
\examples{
nwos_proportion(weight = wi$WEIGHT, domain = wi$Y_1)
nwos_proportion(weight = wi$WEIGHT, area = wi$AC_WOOD, domain = wi$Y_1)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
