% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_proportion_svl.R
\name{nwos_proportion_svl}
\alias{nwos_proportion_svl}
\title{NWOS Proportions by S / V / L}
\usage{
nwos_proportion_svl_replicates(data.value = own.rsvl.rep, data.total = own.rsvl.tot.rep)
}
\arguments{
\item{data.value}{= own.rsvl.rep}

\item{data.total}{= own.rsvl.tot.rep}
}
\description{
Calculate NWOS Proportions by S / V / L
}
\examples{
??
}
\keyword{nwos}
