% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_proportion_svl_replicates.R
\name{nwos_proportion_svl_replicates}
\alias{nwos_proportion_svl_replicates}
\title{NWOS Proportions by S / V / L for Replicates}
\usage{
nwos_proportion_svl_replicates(rep, data.value = own.rsvl.rep, data.total = own.rsvl.tot.rep)
}
\arguments{
\item{rep}{= 1,}

\item{data.value}{= own.rsvl.rep}

\item{data.total}{= own.rsvl.tot.rep}
}
\description{
Calculate NWOS Proportions by S / V / L for Replicates
}
\examples{
??
}
\keyword{nwos}
