% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_quantile.R
\name{nwos_quantile}
\alias{nwos_quantile}
\title{NWOS Quantile}
\usage{
nwos_quantile(weight, area = 1, domain, variable, prob=c(0.00, 0.25, 0.50, 0.75, 1.00), max.iter=1000)
}
\arguments{
\item{weight}{vector of weights per ownership.}

\item{area}{vector of area (e.g., forest acres) per ownership. Default = 1 (i.e., estimates are in terms of ownerships).}

\item{domain}{vector with 1 indicating inclusion in the domain and 0 otherwise. Default = 1 (i.e., all ownerships are included).}

\item{variable}{vector of variable of interest in terms of the quantile.}

\item{prob}{vector of probabilities for quantiles. Defualt is 0.00, 0.25, 0.50, 0.75, 1.00 (i.e., quartiles).}

\item{max.iter}{maximum number of iterations. Default = 1000.}
}
\description{
This function calculates quantiles for NWOS estimates.
}
\examples{
nwos_quantile(weight = wi$WEIGHT, domain = wi$FFO, variable = wi$AC_WOOD)
nwos_quantile(weight = wi$WEIGHT, area = wi$AC_WOOD, domain = wi$FFO, variable = wi$AC_WOOD)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
