% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_region.R
\name{nwos_region}
\alias{nwos_region}
\title{NWOS Region}
\usage{
nwos_region(svl = svl, data = own, states.ref = STATES_REF)
}
\arguments{
\item{svl}{= svl, data = own, states.ref = STATES_REF}
}
\description{
Calculate national, regional, and subregional summaries
}
\examples{
??
}
\keyword{nwos}
