% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_replicates.R
\name{nwos_replicates}
\alias{nwos_replicates}
\title{NWOS Replicates}
\usage{
nwos_replicates(df, R=5000)
}
\arguments{
\item{R}{number of replicates or bootstraps. Default is 5000.}

\item{df}{data frame of survey data.}
}
\value{
Returns a list containing the original data frame with an INDEX appended and a matrix of replicates.
}
\description{
Create a matrix of This function calculates variances for NWOS statistics using a bootstrapping approach.
}
\details{
This function needs to be run by statev
}
\examples{
load("../../ANALYSIS/SAMPLE_SIZE/DATA/WI_SAMPLE_SIZE_SAMPLES_20180721.RData")
df <- SAMPLE_SIZE_SAMPLES[SAMPLE_SIZE_SAMPLES$STATE_CD == 55 & SAMPLE_SIZE_SAMPLES$SAMPLE==1 & SAMPLE_SIZE_SAMPLES$SIZE==1000, ]
df.replicates <- nwos_replicates(df)
save(df.replicates, file = "data/DF_REPLICATES.RData")
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
