% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_replicates_state.R
\name{nwos_replicates_state}
\alias{nwos_replicates_state}
\title{NWOS Replicates by State}
\usage{
nwos_replicates_state(x, data = po)
}
\value{
??
}
\description{
Generate replicates by state that can be used for NWOS bootstrapping variance estimation.
}
\examples{
??
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
