% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_response_rate.R
\name{nwos_response_rate}
\alias{nwos_response_rate}
\title{nwos_response_rate}
\usage{
nwos_response_rate(x)
}
\arguments{
\item{x}{is a vector containing values of R ('response'), NR ('non-response'), and ENR ('excused non-response' or undeliverable)}
}
\value{
a numeric value
}
\description{
Calculates nwos response rate
}
\examples{
nwos_response_rate(c('R','R','NR','NR','ENR','ENR','ENR'))

}
