% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_response_rate_apply.R
\name{nwos_response_rate_apply}
\alias{nwos_response_rate_apply}
\title{NWOS Replicate Response Rates}
\usage{
nwos_response_rate(index.rep, index, stratum, response)
}
\arguments{
\item{index.rep}{list of observations (i.e., replicates) to include.}

\item{index}{vector used to identify the location of values in the other vectors (e.g., row names).}

\item{stratum}{vector indicating inclusion (1) and exclusion (0) in the stratum of interest. NA's are allowed.}

\item{response}{vector indicating response (1) and non-response (0).}
}
\value{
Response rate in the stratum.
}
\description{
This function calculates response rates for NWOS replicates and is designed to be used with an apply function, such as sapply.
}
\examples{
wi <- tbl_df(read.csv("data/wi.csv")) \%>\% mutate(ROW_NAME = row.names(wi), AC_WOOD = ACRES_FOREST, FFO = if_else(LAND_USE == 1 & OWN_CD == 45 & AC_WOOD >= 1, 1, 0), RESPONSE = if_else(RESPONSE_PROPENSITY >= 0.5, 1, 0), RESPONSE = if_else(is.na(RESPONSE_PROPENSITY), 0, RESPONSE))
WI_FFO_RR_REP <- sapply(WI_REPLICATES, nwos_response_rate_apply, index = wi$ROW_NAME, stratum = wi$FFO, response = wi$RESPONSE)
WI_FFO_RR_REP
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
