% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_response_rate_replicates.R
\name{nwos_response_rate_replicates}
\alias{nwos_response_rate_replicates}
\title{NWOS Replicate Response Rates}
\usage{
nwos_response_rate(index, data, stratum.name, response.name)
}
\arguments{
\item{index}{list of observations (i.e., replicates) in data to include.}

\item{data}{data frame containing stratum variable.}

\item{stratum.name}{the name of the variable in data indicating inclusion (1) and exclusion (0) in the stratum of interest}

\item{response.name}{name of the variable in data indicating response (1) and non-response (0).}
}
\value{
Response rate in the stratum.
}
\description{
This function calculates response rates for NWOS replicates and is designed to be used with an apply function, such as sapply.
}
\details{
This function needs to be run by state.
}
\examples{
wi <- wi \%>\% mutate(RESPONSE = if_else(RESPONSE_PROPENSITY >= 0.5, 1, 0))
nwos_response_rate_replicates(index = WI_REPLICATES[[1]], data = wi, stratum.name = "FFO", response.name = "RESPONSE")
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
