% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_stratum_area.R
\name{nwos_stratum_area}
\alias{nwos_stratum_area}
\title{NWOS Stratum Area}
\usage{
nwos_stratum_area(stratum, point.count, state.area)
}
\arguments{
\item{state.area}{vector of total land area in state. It needs to have the same length as state.list and be in the
same order.}

\item{df}{data frame containing, at a minimum, the following variables: STATE_CD, LAND_USE, OWN_CD, and POINT_COUNT}

\item{state.list}{vector of state codes to be evaluated.}

\item{land.use.list}{vector of land uses to be evaluated. The defaualt value is 1 (forest).}

\item{own.cd.list}{vector of ownership codes to be evaluated. The default value if 45 (family).}
}
\value{
data frame of areas in state/land use/ownership strata
}
\description{
This function estimates the stratum area.
}
\examples{
load("data/nwos_data_sample.RData")
data <- NWOS_DATA_SAMPLE[NWOS_DATA_SAMPLE$SAMPLE==1,]
nwos_stratum_area()
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
