% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_stratum_area_replicates.R
\name{nwos_stratum_area_replicates}
\alias{nwos_stratum_area_replicates}
\title{NWOS Replicate Stratum Areas}
\usage{
nwos_stratum_area_replicates(index, data, stratum.name, state.area)
}
\arguments{
\item{index}{list of observations (i.e., replicates) in data to include.}

\item{data}{data frame containing stratum variable.}

\item{stratum.name}{the name of the variable in data indicating inclusion (1) and exclusion (0) in the stratum of interest}

\item{state.area}{total land area in the state of interest.}

\item{point.count}{the number of sample points associated with each observation.}
}
\value{
Area of forestland in the stratum.
}
\description{
This function estimates NWOS stratum areas for NWOS replicates and is designed to be used with an apply function, such as sapply.
}
\details{
This function needs to be run by state.
point.count is automatically set to 1 since the replicates are an expanded list (one observation per sample point).
}
\examples{
wi <- wi \%>\% mutate(FFO = if_else(LAND_USE == 1 & OWN_CD == 45 & AC_WOOD >= 1, 1, 0)) # Add stratum variable, FFO
WI_FFO_AREA_REP <- sapply(WI_REPLICATES, nwos_stratum_area_replicates, data = wi, stratum.name = "FFO", state.area = 33898733)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
