% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_stratum_area_replicates_state.R,
%   R/nwos_weights_stratum_area_replicates_state.R
\name{nwos_stratum_area_replicates_state}
\alias{nwos_stratum_area_replicates_state}
\title{NWOS Stratum Areas for Replicates by State}
\usage{
nwos_stratum_area_apply(index.rep, index, stratum, state.area)

nwos_stratum_area_apply(index.rep, index, stratum, state.area)
}
\arguments{
\item{data}{= po}

\item{s.name}{= "FFO"}
}
\value{
Area of forestland in the stratum.

Area of forestland in the stratum.
}
\description{
This function estimates NWOS stratum areas for NWOS replicates and is designed to be used with an apply function, such as sapply.

This function estimates NWOS stratum areas for NWOS replicates and is designed to be used with an apply function, such as sapply.
}
\details{
point.count in nwos_stratum_area is  set to 1 since the replicates are an expanded list (one observation per sample point).

point.count in nwos_stratum_area is  set to 1 since the replicates are an expanded list (one observation per sample point).
}
\examples{
wi <- tbl_df(read.csv("data/wi.csv")) \%>\% mutate(ROW_NAME = row.names(wi), AC_WOOD = ACRES_FOREST, FFO = if_else(LAND_USE == 1 & OWN_CD == 45 & AC_WOOD >= 1, 1, 0))
WI_REPLICATES <- nwos_replicates(index = row.names(wi), point.count = wi$POINT_COUNT, R = 100)
WI_FFO_AREA_REP <- sapply(WI_REPLICATES, nwos_stratum_area_apply, index = wi$ROW_NAME, stratum = wi$FFO, state.area = 33898733)
WI_FFO_AREA_REP
wi <- tbl_df(read.csv("data/wi.csv")) \%>\% mutate(ROW_NAME = row.names(wi), AC_WOOD = ACRES_FOREST, FFO = if_else(LAND_USE == 1 & OWN_CD == 45 & AC_WOOD >= 1, 1, 0))
WI_REPLICATES <- nwos_replicates(index = row.names(wi), point.count = wi$POINT_COUNT, R = 100)
WI_FFO_AREA_REP <- sapply(WI_REPLICATES, nwos_stratum_area_apply, index = wi$ROW_NAME, stratum = wi$FFO, state.area = 33898733)
WI_FFO_AREA_REP
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.

Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
