% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_stratum_area_state.R
\name{nwos_stratum_area_state}
\alias{nwos_stratum_area_state}
\title{NWOS Stratum Area by State}
\usage{
nwos_stratum_area_state(state, data = po, s.name = "FFO", data.area = land.area, ac.name = "ACRES")
}
\arguments{
\item{data}{= po}

\item{s.name}{= "FFO"}

\item{data.area}{= land.area}

\item{ac.name}{= "ACRES"}
}
\value{
Area (of forestland) in the stratum.
}
\description{
This function estimates NWOS stratum areas by state.
}
\examples{
??
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
