% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_svl.R
\name{nwos_svl}
\alias{nwos_svl}
\title{NWOS States, Variables, and Levels}
\usage{
nwos_svl(data = QUEST_WIDE)
}
\arguments{
\item{data}{= QUEST_WIDE}
}
\description{
Generate List of States, Variables, and Levels used for NWOS functions
}
\examples{
??
}
\keyword{nwos}
