% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table.R
\name{nwos_table}
\alias{nwos_table}
\title{nwos_table}
\usage{
nwos_table(
  tab.num = 12.1,
  data = table.data,
  stratum = stratum.name,
  domain = domain.name,
  yr = year,
  yr.range = year.range
)
}
\description{
nwos_table
}
\details{
For area and cooperation rate tables see ...
nwos_table()
nwos_table_area()
nwos_table_coop()
}
