% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table_area.R
\name{nwos_table_area}
\alias{nwos_table_area}
\title{nwos_table}
\usage{
nwos_table_area(
  geo.abb,
  data = AREA,
  yr = year,
  yr.range = year.range,
  ref.tab = REF_TABLE,
  ref.geo = REF_GEO
)
}
\description{
Create the body of an NWOS core, area, or cooperataion rate table
}
\details{
For area and cooperation rate tables see ...
geo.abb = "AL"
data = AREA
yr = year
yr.range = year.range
ref.tab = REF_TABLE
ref.geo = REF_GEO
nwos_table()
nwos_table_area(geo.abb = "AL")
nwos_table_coop()
}
