% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table_data.R
\name{nwos_table_data}
\alias{nwos_table_data}
\title{nwos_tables_data}
\usage{
nwos_table_data(
  data = ESTIMATES_WIDE,
  units = "thousands",
  ref.geo = REF_GEO,
  ref.tab = REF_TABLE,
  ref.var = REF_VARIABLE,
  ref.lab = REF_LABEL
)
}
\description{
Create the body of an NWOS core, area, or cooperataion rate table
}
\details{
For area and cooperation rate tables see ...
nwos_table_data()
}
