% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table_make_estimates_wide.R
\name{nwos_table_make_estimates_wide}
\alias{nwos_table_make_estimates_wide}
\title{nwos_table_set}
\usage{
nwos_table_make_estimates_wide(estimates = QUEST_EST, domain)
}
\arguments{
\item{TAB_TYPE}{table type. AREA = area by ownership category. COOP = cooperation rate. QUEST (Default) = questionnaire content.}

\item{POP}{. Default = "Family".}

\item{DOMAIN}{= NA}

\item{YEARS}{Default = "2017-2018"}
}
\description{
Create the body of an NWOS table
}
\details{
For area and cooperation rate tables see ...
nwos_table_body()
nwos_table_body_area()
nwos_table_body_coop()
}
\examples{
nwos_table_make_estimates_wide(quest.file = "INPUTS/ESTIMATES/NWOS_2018_FFO_ESTIMATES_20200221.RDS",
domain = "AC_WOOD >= 1")

}
