% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table_make_wide.R
\name{nwos_table_make_wide}
\alias{nwos_table_make_wide}
\title{nwos_table_make_wide}
\usage{
nwos_table_make_wide(data)
}
\description{
Create the body of an NWOS table
}
\examples{
x <- nwos_table_make_wide(data = readRDS("INPUTS/ESTIMATES/NWOS_2018_FFO_TENPLUS_US.RDS"))

}
