% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table_total.R
\name{nwos_table_total}
\alias{nwos_table_total}
\title{nwos_table_total}
\usage{
nwos_table_total(geo.abb, data = QUEST_EST, ref.geo = REF_GEO,
  ref.tab = REF_TABLE, stratum.abb = STRATUM_ABB,
  stratum.name = STRATUM_NAME, domain.abb = DOMAIN_ABB,
  domain.name = DOMAIN_NAME, year = YEAR, year.range = YEAR_RANGE)
}
\description{
geo.abb = "US"
data = ESTIMATES_WIDE
ref.geo = REF_GEO
ref.tab = REF_TABLE
stratum = stratum.name
domain = domain.name
yr = year
yr.range = year.range
nwos_table_total("US")
}
