% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_tables_raw.R
\name{nwos_tables_raw}
\alias{nwos_tables_raw}
\title{nwos_tables_raw}
\usage{
nwos_tables_raw(
  area_data = NA,
  coop_data = NA,
  quest_data = QUEST_EST,
  quest_tab_meta_data = REF_TABLE
)
}
\description{
Create the body of an NWOS core, area, or cooperataion rate table
}
\details{
For area and cooperation rate tables see ...
nwos_table_raw()
}
