% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_total.R
\name{nwos_total}
\alias{nwos_total}
\title{NWOS Total}
\usage{
nwos_total(weight, area = 1, domain = 1, variable = 1)
}
\arguments{
\item{weight}{vector of weights per ownership.}

\item{area}{vector of area (e.g., forest acres) per ownership.}

\item{domain}{vector with 1 indicating inclusion in the domain and 0 otherwise.}

\item{variable}{vector of variable of interest.}
}
\description{
Estimators used to calculate NWOS total.
}
\details{
If area is set to 1, then the esitmates are in terms of ownerships.
Variable can be binary or continuous. If variable is binary, the value returned is the number of ownerships or acreage in the domain of interest with that attribute. If variable is continuous, the value returned is the variable total in terms of ownerships or acreage.
}
\examples{
WI_FFO_OWN_TOTAL <- nwos_total(weight = wi$WEIGHT)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
