% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_total_apply.R
\name{nwos_total_apply}
\alias{nwos_total_apply}
\title{NWOS Total Apply}
\usage{
nwos_total_appl(index, weight, area, domain, variable)
}
\arguments{
\item{index}{list of vectors indicating which data to include. There needs to be an element with the same name as the replicate being assessed.}

\item{weight}{list of vectors of weights. There needs to be an element with the same name as the replicate being assessed.}

\item{area}{vector of area (e.g., forest acres) per ownership.}

\item{domain}{vector with 1 indicating inclusion in the domain and 0 otherwise.}

\item{variable}{vector of variable of interest.}

\item{replicate}{an atomic value indicating which replicate to use.}
}
\description{
A version of the nwos_total function that is intended to be used with an apply function, typically for estimating sampling errors.
}
\details{
If index is set to 1:length(x), the results are equivalent to nwos_total.
Default values for area, domain, and variable are not set, as they are for nwos_total, in order to reduce computation time.
}
\examples{
NEED TO ADD
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
