% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_total_replicates.R
\name{nwos_total_replicates}
\alias{nwos_total_replicates}
\title{NWOS Replicate Totals}
\usage{
nwos_total_replicates <- function(r, index.rep, data, weight.rep, owner.area.name = "OWNER", domain.name = "FFO", variable.name = "FFO")
}
\arguments{
\item{r}{vector of replicates. See details below.}

\item{index.rep}{list of observations (i.e., replicates) in data to include.}

\item{data}{data frame containing stratum variable.}

\item{owner.area.name}{nae of a variable in data of the area (of forestland) owned by each ownership. Default = "OWNER" (i.e., 1, which is used for estimating number of ownerships).}

\item{domain.name}{name of the variable in data indicating inclusion (1) and exclusion (0) in the domain of interest. Default = "FFO".}

\item{variable.name}{name of the variable of interest in data. Default = "FFO". See details below.}
}
\description{
Estimators used to calculate NWOS totals for replicates. This is typically used with an apply function.
}
\details{
r it typcially set using 1:length(WI_REPLICATES)
For estiamtes in terms ownerships, a dummy variable (e.g., OWNER=1) needs to be added.
Variable can be binary or continuous. If variable is binary, the value returned is the number of ownerships or acreage in the domain of interest with that attribute. If variable is continuous, the value returned is the variable total in terms of ownerships or acreage.
}
\examples{
wi <- wi \%>\% mutate(OWNER = 1)
WI_FFO_OWN_TOTAL_REP <- sapply(1:length(WI_REPLICATES), nwos_total_replicates, index.rep = WI_REPLICATES, data = wi,
weight.rep = WI_FFO_WEIGHTS_REP, owner.area.name = "OWNER",
domain.name = "FFO", variable.name = "FFO")
var(WI_FFO_OWN_TOTAL_REP)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
