% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_total_svl_replicates.R
\name{nwos_total_svl_replicates}
\alias{nwos_total_svl_replicates}
\title{NWOS Total by State, Variable, and Level for Replicates}
\usage{
nwos_total_svl(states, variables, levels, data = QUEST, data.area = NA, data.weights = "WEIGHT")
}
\arguments{
\item{states}{= "1"}

\item{variables}{= "HOME"}

\item{levels}{= "1",}

\item{data}{= QUEST_WIDE}

\item{data.area}{= NA}

\item{data.weights}{= "WEIGHT"}
}
\description{
Estimator used to calculate NWOS totals by state, variable and level.
}
\examples{
??
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
