% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_urban_table.R
\name{nwos_urban_table}
\alias{nwos_urban_table}
\title{nwos_table}
\usage{
nwos_urban_table(tab.num, data = table.data,
  data.coop = coop.table.data, stratum.abb = STRATUM_ABB,
  stratum.name = STRATUM_NAME, domain.abb = DOMAIN_ABB,
  domain.name = DOMAIN_NAME, year = YEAR)
}
\description{
nwos_table
}
\details{
For area and cooperation rate tables see ...
tab.num = 3
nwos_table()
nwos_table_area()
nwos_table_coop()
}
