% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_variance.R
\name{nwos_variance}
\alias{nwos_variance}
\title{NWOS Variance}
\usage{
nwosVar(point.count, domain, y=1, area, response=1,
stratum.area, units="ownerships", stat="total", R=100)
}
\arguments{
\item{stat}{statistic of interest. Permissable values are "total", "mean" or "proportion". Default is "total".}

\item{units}{units of analysis. Permissable values are "ownerships" or "area". Default is "ownerships".}

\item{data}{data frame of survey data. See details below for required variables.}

\item{R}{number of replicates or bootstraps. Default is 100.}
}
\description{
This function calculates variances for NWOS statistics using a bootstrapping approach.
}
\examples{
x <- nwos_variances(df=sample.i.j, state.list=55, state.area=LAND_AREA, R=10)
x$VARIANCES
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
