% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_weights.R
\name{nwos_weights}
\alias{nwos_weights}
\title{NWOS Weights}
\usage{
nwosWeights(point.count, area, domain, stratum.area, response.rates)
}
\arguments{
\item{stratum.area}{area for stratum}

\item{response.rate}{response rate for stratum. Default value = 1.}

\item{point.count}{vector of counts of sample points per ownership.}

\item{area}{vector of area (e.g., forest acres) per ownership.}
}
\value{
vector of weights
}
\description{
This function returns the calculated weights based on the NWOS sample design.
}
\examples{
load("data/nwos_data_sample.RData")
data <- NWOS_DATA_SAMPLE[NWOS_DATA_SAMPLE$SAMPLE==1,]
data$w <- nwosWeights(data$POINT_COUNT, data$ACRES_FOREST, 1000)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
