% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_weights_replicates.R
\name{nwos_weights_replicates}
\alias{nwos_weights_replicates}
\title{NWOS Replicate Weights}
\usage{
nwos_weights_replicates(r, index.rep, data, stratum.name = "FFO", point.count.name = "POINT_COUNT",
response.name = "RESPONSE", owner.area.name = "AC_WOOD", stratum.area.rep, response.rate.rep)
}
\arguments{
\item{r}{vector of replicates. See details below.}

\item{index.rep}{list of observations (i.e., replicates) in data to include.}

\item{data}{data frame containing stratum variable.}

\item{stratum.name}{name of the variable in data indicating inclusion (1) and exclusion (0) in the stratum of interest. Default = "FFO".}

\item{point.count.name}{name of a variable in data of the number of sample points associated with each observation.. Default = "POINT_COUNT",}

\item{response.name}{name of a variable in data indicating response (1) and non-response (0). Default = "RESPONSE".}

\item{owner.area.name}{nae of a variable in data of the area (of forestland) owned by each ownership. Default = "AC_WOOD".}

\item{stratum.area.rep}{vector of area (of forestland) in the stratum of interest for each replicate.}

\item{response.rate.rep}{vector of response rates for each replicate for the stratum of interest.}
}
\value{
vector of weights
}
\description{
This function returns the calculated weights based on the NWOS sample design and is designed to be used with an apply function, such as sapply.
}
\details{
r it typcially set using 1:length(WI_REPLICATES)
index.rep differs from index in other nwos functions in that these includes multiple lists (i.e., all of the replicates, not just one).
index.rep, stratum.area.rep, and response.rate.rep are assumed to be listed in the same order.
}
\examples{
WI_FFO_WEIGHTS_REP <- lapply(1:length(WI_REPLICATES), nwos_weights_replicates,
index.rep = WI_REPLICATES, data = wi,
stratum.area.rep = WI_FFO_AREA_REP, response.rate.rep = WI_FFO_RR_REP)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
