% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_weights_replicates_state.R
\name{nwos_weights_replicates_state}
\alias{nwos_weights_replicates_state}
\title{NWOS Weights for Replicates by State}
\usage{
nwos_weights_replicates(r, index.rep, index, stratum, response, area, stratum.area, stratum.area.correction = stratum.area, response.rate)
}
\arguments{
\item{data}{= po}

\item{s.name}{= "FFO"}

\item{r.name}{= "RESPONSE"}
}
\value{
list of vector of weights
}
\description{
This function returns the calculated weights based on the NWOS sample design and is designed to be used with an apply function, such as lapply.
}
\examples{
??
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
