% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_weights_state.R
\name{nwos_weights_state}
\alias{nwos_weights_state}
\title{NWOS Weights by State}
\usage{
nwos_weights(stratum, point.count, response, area, stratum.area, stratum.area.correction = stratum.area, response.rate)
}
\arguments{
\item{state}{state}

\item{data}{= po}
}
\value{
vector of weights
}
\description{
This function returns the calculated weights based on the NWOS sample design.
}
\examples{
??
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
