% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_wide_metadata.R
\name{nwos_wide_metadata}
\alias{nwos_wide_metadata}
\title{nwos_wide_metadata}
\usage{
nwos_wide_metadata(x)
}
\arguments{
\item{x}{is a nwos.object}
}
\value{
a data.frame
}
\description{
Creates metadata table for NWOS data in wide format
}
\examples{
nwos_wide_metadata(foo)

}
