% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_data.R
\name{nwosData}
\alias{nwosData}
\title{NWOS Data}
\usage{
nwosData()
}
\description{
This function creates a dataset that can be used to test the NWOS functions.
}
\examples{
NWOS_DATA <- nwosData()
save(NWOS_DATA, file="data/nwos_data.RData")
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
