% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_data_sample.R
\name{nwosDataSample}
\alias{nwosDataSample}
\title{NWOS Data Sample}
\usage{
nwosDataSample(df,n,s)
}
\arguments{
\item{df}{dataframe with NWOS structure (see Description file)}

\item{n}{number of samples per sample iteration}

\item{s}{number of sample iterations}
}
\value{
Dataframe for sample
}
\description{
This function generates one or more samples of NWOS data using a probability proportional to size, with replacement, design.
}
\examples{
load("data/nwos_data.RData")
NWOS_DATA_SAMPLE <- nwosDataSample(NWOS_DATA, n=25, s=1000)
save(NWOS_DATA_SAMPLE, file="data/nwos_data_sample.RData")
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
