% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWOSdelete.R
\name{NWOSdelete}
\alias{NWOSdelete}
\title{NWOSdelete}
\usage{
NWOSdelete(tables)
}
\arguments{
\item{tables}{is a character vector containing names of vectors currently in the working environment; each vector contains CNs to be deleted from the corresponding table.}
}
\description{
Deletes records from the NWOS database
}
\details{
This function must be run inside an open RODBC channel (32-bit only) connected to the USFS FIA production database through a user with write permissions.
}
\examples{
NWOSdelete(tables=c('RESPONSE_DELETE'))

}
