% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWOSdelete64.R
\name{NWOSdelete64}
\alias{NWOSdelete64}
\title{NWOSdelete64}
\usage{
NWOSdelete64(tables)
}
\arguments{
\item{tables}{is a character vector containing names of vectors currently in the working environment; each vector contains CNs to be deleted from the corresponding table.}
}
\description{
Deletes records from the NWOS database from a 64-bit R session
}
\details{
This function must be run on a machine with an ODBC connection (32-bit) to the USFS FIA production database through a user with read permissions.
}
\examples{
NWOSdelete64(tables=c('RESPONSE_DELETE'))

}
