% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWOSload.R
\name{NWOSload}
\alias{NWOSload}
\title{NWOSload}
\usage{
NWOSload(tables)
}
\arguments{
\item{tables}{is a character vector containing names of update tables currently in the working environment.}
}
\description{
Loads update tables to the NWOS database
}
\details{
This function must be run inside an open RODBC channel (32-bit only) connected to the USFS FIA production database through a user with write permissions.
}
\examples{
NWOSload(tables=c('SAMPLE_UPDATE','OWNER_UPDATE'))

}
