% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_mean.R
\name{nwosMean}
\alias{nwosMean}
\title{NWOS Mean}
\usage{
nwosMean(weight, point.count, domain, y=1, area, units="ownerships")
}
\arguments{
\item{weight}{weight for each observation.}

\item{y}{variable of interest. Set to 1 if interested in basic ownership or area totals. Default is 1.}

\item{area}{vector of areas of forest land owned by sampled ownerships. NAs are permissable.}

\item{domain}{variable indicating whether ownership is in the domain of interest.}

\item{units}{units of analysis. Permissable values are "ownerships" or "area". Default is "ownerships".}
}
\description{
This function calculates totals for the NWOS.
}
\examples{
load("data/nwos_data_sample.RData")
df <- NWOS_DATA_SAMPLE[NWOS_DATA_SAMPLE$SAMPLE==1,]
df$WEIGHT <- nwosWeights(df$POINT_COUNT, df$ACRES_FOREST, 1000)
# Mean acreage by ownership
nwosMean(weight=df$WEIGHT, y=df$ACRES_FOREST, var=F)
# Mean acreage by area
nwosMean(weight=df$WEIGHT, y=df$ACRES_FOREST, area=df$ACRES_FOREST, units="area", var=F)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
