% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_proportion.R
\name{nwosProportion}
\alias{nwosProportion}
\title{NWOS Proportion}
\usage{
nwosProportion(weight, point.count, domain, y, area, units="ownerships")
}
\arguments{
\item{weight}{weight for each observation.}

\item{point.count}{vector of number of sample points. Needs to sum to total number of sample points across all land uses and ownership classes.}

\item{domain}{variable indicating whether ownership is in the domain of interest.}

\item{y}{variable of interest. Set to 1 if interested in basic ownership or area totals.}

\item{area}{vector of areas of forest land owned by sampled ownerships. NAs are permissable.}

\item{units}{units of analysis. Permissable values are "ownerships" or "area". Default is "ownerships".}
}
\description{
This function calculates totals for the NWOS.
}
\examples{
load("data/nwos_sample_data.RData")
nwos.sample.data$domain <- ifelse(nwos.sample.data$owner.class=="FamilyForest", 1, 0)
nwos.sample.data$response <- ifelse(nwos.sample.data$owner.class=="FamilyForest",
as.numeric(as.character(nwos.sample.data$response)), NA)
nwos.sample.data$weights <- nwosWeights(point.count=nwos.sample.data$point.count,
area=nwos.sample.data$area,
domain=nwos.sample.data$domain,
stratum.area=35198019,
response.rate=sample.response.rate)
# Proportion of ownership with y_1=1
nwos.sample.data$y_1 <- as.numeric(as.character(nwos.sample.data$y_1))
nwosProportion(weight=nwos.sample.data$weight,
point.count=nwos.sample.data$point.count,
domain=nwos.sample.data$domain,
y=nwos.sample.data$y_1,
area=nwos.sample.data$area,
units="ownerships")
# Proportion of area with y_1=1
nwosProportion(weight=nwos.sample.data$weight,
point.count=nwos.sample.data$point.count,
domain=nwos.sample.data$domain,
y=nwos.sample.data$y_1,
area=nwos.sample.data$area,
units="area")
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
