% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_replication.R
\name{nwosReplication}
\alias{nwosReplication}
\title{NWOS Replication}
\usage{
nwosMean(weight, point.count, domain, y=1, area, units="ownerships")
}
\arguments{
\item{df}{is a data frame}

\item{r}{is the number of replicates for each observation}
}
\description{
This function replicates the rows of a dataframe based on plot count.
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
