% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_response_rate.R
\name{nwosResponseRate}
\alias{nwosResponseRate}
\title{NWOS Response Rate}
\usage{
nwosResponseRate(point.count, response)
}
\arguments{
\item{point.count}{vector of counts of sample points per ownership/record.}

\item{response}{vector with 1 indicating response and 0 nonresponse.}
}
\value{
response rate
}
\description{
This function calculates response rate for the NWOS.
}
\examples{
load("data/nwos_data_sample.RData")
data <- NWOS_DATA_SAMPLE[NWOS_DATA_SAMPLE$SAMPLE==1,]
nwosResponseRate(data$POINT_COUNT, data$RESPONSE)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
