% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_response_rate.R
\name{nwosResponseRates}
\alias{nwosResponseRates}
\title{NWOS Response Rates}
\usage{
nwosResponseRates(point.count, response)
}
\arguments{
\item{point.count}{vector of number of sample points. Needs to sum to total number of sample points across all land uses and ownership classes.}

\item{response}{Indicator variable for whether an ownership responded. 1=Yes and 0=No.}
}
\value{
response rate
}
\description{
This function calculates response rates for the NWOS by stratum and domain.
}
\examples{
## Calculate response rate for family forest ownerships
load("data/nwos_sample_data.RData")
nwos.sample.data$response <- ifelse(nwos.sample.data$owner.class=="FamilyForest",
as.numeric(as.character(nwos.sample.data$response)), NA)
nwosResponseRates(nwos.sample.data$point.count,nwos.sample.data$response)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
