% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_weights_trimmed.R
\name{nwosTrimWeights}
\alias{nwosTrimWeights}
\title{NWOS Trim Weights}
\usage{
nwosTrimWeights(weights, method = "iqr_1.5")
}
\arguments{
\item{method}{method to be used for trimming. The only current method implemented is "iqr_1.5".}

\item{weight}{weight for each observation.}
}
\description{
This function calculates trimmed weights for the NWOS.
}
\examples{
# Calculate trimmed weights for family forest ownerships
load("data/nwos_sample_data.RData")
nwos.sample.data$domain <- ifelse(nwos.sample.data$owner.class=="FamilyForest", 1, 0)
nwos.sample.data$response <- ifelse(nwos.sample.data$owner.class=="FamilyForest",
as.numeric(as.character(nwos.sample.data$response)), NA)
sample.response.rate <- nwosResponseRates(nwos.sample.data$point.count,nwos.sample.data$response)
nwos.sample.data$weights <- nwosWeights(point.count=nwos.sample.data$point.count,
area=nwos.sample.data$area,
domain=nwos.sample.data$domain,
stratum.area=35198019,
response.rate=sample.response.rate)
nwos.sample.data$weights.trimmed <- nwosTrimWeights(nwos.sample.data$weights)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
