% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nxtcn.R
\name{nxtcn}
\alias{nxtcn}
\title{nxtcn}
\usage{
nxtcn(x)
}
\arguments{
\item{x}{is a character vector of NWOS CNs}
}
\value{
a numeric vector
}
\description{
Identifies the numerically subsequent NWOS CN number, for appending new records to a table
}
\examples{
maxcn(c('SAM00001','SAM04567'))

}
