% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasten.R
\name{pasten}
\alias{pasten}
\title{pasten}
\usage{
pasten(..., sep = " ")
}
\arguments{
\item{x}{is a series of string vectors}
}
\value{
a string vector
}
\description{
Paste function where nulls equal ""
}
\examples{
pasten('an unbroken',NA,'string')

}
