% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qumx.R
\name{qumx}
\alias{qumx}
\title{qumx}
\usage{
qumx(x)
}
\arguments{
\item{x}{is a character string containing the name of an NWOS table (without schema named appended)}
}
\value{
a character string
}
\description{
Generates the appropriate SQL query for selecting the most recent record in a table (the record with the numerically highest CN number) in the NWOS DB
}
\examples{
qumx('SAMPLE')

}
