% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_nwos_plots_object.R
\name{register_nwos_plots_object}
\alias{register_nwos_plots_object}
\title{register_nwos_plots_object}
\usage{
register_nwos_plots_object()
}
\description{
Registers S4 class 'nwos.plots' in working environment
}
\examples{
register_nwos_plots_object(foo)

}
