% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlQuery64.R
\name{sqlQuery64}
\alias{sqlQuery64}
\title{sqlQuery64}
\usage{
sqlQuery64(query)
}
\arguments{
\item{query}{is a string containing the text of the SQL query to be sent to the NWOS DB}
}
\value{
a dataframe containing the return from the SQL query
}
\description{
Wrapper function for RODBC:sqlQuery to send queries to the NWOS database from a 64-bit R session
}
\details{
This function must be run on a machine with an ODBC connection (32-bit) to the USFS FIA production database through a user with read permissions.
}
\examples{
sqlQuery64(query='SELECT * FROM FS_NWOS.FIELDS')

}
