% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableformat.R
\name{tableformat}
\alias{tableformat}
\title{tableformat}
\usage{
tableformat(type = "update")
}
\arguments{
\item{type}{a character vector equal to 'full' or 'update'.}
}
\description{
Formats NWOS DB tables, both full tables and update tables
}
\details{
This function must be run on a machine with an ODBC connection (32-bit) to the USFS FIA production database through a user with read permissions.
}
\examples{
tableformat(type='full')

}
