% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjur.R
\name{tjur}
\alias{tjur}
\title{tjur}
\usage{
tjur(data, variable = "RESPONSE_CD", pred.prob = "PRED")
}
\arguments{
\item{data}{is a data frame}

\item{variable}{is the name of the binary response variable}

\item{pred.prob}{is the predicted probability}
}
\description{
Calculate Tjur's Statistic
}
\details{
Calculates Tjur's statistic for predictions from a binary model
}
\examples{
tjur(data = model.data, variable = "RESPONSE_CD", pred.prob = "PRED")

}
