% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{trim}
\usage{
trim(x)
}
\arguments{
\item{x}{is a character vector}
}
\value{
a character vector with leading and trailing spaces removed
}
\description{
Trims leading and trailing whitespace
}
\examples{
trim(" foo ")

}
