% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trlow.R
\name{trlow}
\alias{trlow}
\title{trlow}
\usage{
trlow(x)
}
\arguments{
\item{x}{is a character vector}
}
\value{
an lowercase character vector with leading and trailing spaces removed
}
\description{
Trims and sets to lowercase
}
\examples{
trlow(" foo ")

}
