% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trup.R
\name{trup}
\alias{trup}
\title{trup}
\usage{
trup(x)
}
\arguments{
\item{x}{is a character vector}
}
\value{
an uppercase character vector with leading and trailing spaces removed
}
\description{
Trims and sets to uppercase
}
\examples{
trup(" foo ")

}
