% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_all.R
\name{unique.all}
\alias{unique.all}
\title{unique.all}
\usage{
\method{unique}{all}(x)
}
\arguments{
\item{x}{is a vector}
}
\value{
a logical value
}
\description{
Determines whether a vector contains only unique, non-null values
}
\examples{
unique.all(c("A","A","B"))

}
