% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwos_plots_intensified.R
\name{get_nwos_plots_intensified}
\alias{get_nwos_plots_intensified}
\title{get_nwos_plots_intensified}
\usage{
get_nwos_plots_intensified(
  cycle = "2018",
  study = "base intensified",
  states = NA,
  yrs = NA,
  strict.intensification = FALSE
)
}
\arguments{
\item{cycle}{is a string containing the NWOS cycle desired.}

\item{study}{is a string containing the NWOS study desired.}

\item{states}{is a string containing the NWOS state desired.}

\item{yrs}{is a logical value determining whether ONLY intensified plots should be selected.}
}
\value{
an nwos.plots object
}
\description{
Extracts the full plot list for a given NWOS cycle/study, classed to STRATA
}
\details{
This function must be run on a machine with an ODBC connection to the USFS FIA production database through a user with read permissions.
}
\examples{
get_nwos_plots(cycle='2018',study='base intensified',states="26",yrs=2017:2018,strict.intensification=FALSE)

}
