% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nwos_plots_urban.R
\name{get_nwos_plots_urban}
\alias{get_nwos_plots_urban}
\title{get_nwos_plots_urban}
\usage{
get_nwos_plots_urban(cycle = "2023", study = "urban", city = "HOUS")
}
\arguments{
\item{cycle}{is a string containing the NWOS cycle desired.}

\item{study}{is a string containing the NWOS study desired.}

\item{city}{is a string containing the NWOS city code desired.}
}
\value{
an nwos.plots object
}
\description{
Extracts the full plot list for a given NWOS cycle/study, classed to STRATA
}
\details{
This function must be run on a machine with an ODBC connection to the USFS FIA production database through a user with read permissions.
}
\examples{
get_nwos_plots_urban(cycle='2023',study='urban',city='HOUS')

}
