% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_cooperation_rate.R
\name{nwos_cooperation_rate}
\alias{nwos_cooperation_rate}
\title{nwos_cooperation_rate}
\usage{
nwos_cooperation_rate(x)
}
\arguments{
\item{x}{is a vector containing values of R ('response'), NR ('non-response'), and ENR ('excused non-response' or undeliverable)}
}
\value{
a numeric value
}
\description{
Calculates nwos cooperation rate
}
\examples{
nwos_cooperation_rate(c('R','R','NR','NR','ENR','ENR','ENR'))

}
