% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates.R
\name{nwos_estimates}
\alias{nwos_estimates}
\title{NWOS Estimates Total}
\usage{
nwos_total_svl(state, variable, level, data = QUEST, area = NA, weights = "WEIGHT")
}
\description{
Estimator used to calculate NWOS totals by state, variable and level.
}
\examples{
quest.list = QUEST_LIST
variable.list = VARIABLE_LIST
geo.list = GEO_LIST
rep = "0"
rep.list = REPLICATE_LIST
stratum = "FFO"
domain = "ONEPLUS"
start.time <- Sys.time()
x <- nwos_estimates(geo.list = GEO_LIST[1,], rep = (0:100))
Sys.time() - start.time
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
