% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_add_tele.R
\name{nwos_estimates_add_tele}
\alias{nwos_estimates_add_tele}
\title{Add TELE Variables to NWOS Dataset}
\usage{
nwos_estimates_add_tele(x, data = QUEST_WIDE)
}
\arguments{
\item{x}{list number. Only applicable if is data is a list of data frames, instead of a single data frame. This used mainly for apply functions.}

\item{data}{data frame or list of data frames}
}
\description{
Add variables to an NWOS dataframe
}
\details{
TELE_ATT: 1= Woodland retreat; 2 = Workign the land; 3 = Supplemental income; 4 = Uninvolved
TELE_PRIME: 1= Model; 2 = Prime; 3 = Defector; 4 = Write-off
}
\examples{
nwos_estimates_add_tele(x = 1 , data = QUEST_LIST)
}
\keyword{nwos}
