% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_geo.R
\name{nwos_estimates_geo}
\alias{nwos_estimates_geo}
\title{nwos_estimates_geo}
\usage{
nwos_estimates_geo(
  domain = "AC_WOOD >= 10",
  geo = GEO,
  cond.status = 1,
  own = 45,
  quest = QUEST,
  plot = PLOTS,
  imputations = 1:5
)
}
\arguments{
\item{domain}{= "AC_WOOD >= 10",}

\item{geo}{= GEO,}

\item{cond.status}{= 1,}

\item{own}{= 45,}

\item{quest}{= QUEST,}

\item{plot}{= PLOTS,}

\item{imputations}{= 1:5}
}
\value{
data frame
}
\description{
Generate estimates by geography
}
\examples{
nwos_estimates_geo()

}
