% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_geo_imp.R
\name{nwos_estimates_geo_imp}
\alias{nwos_estimates_geo_imp}
\title{nwos_estimates_geo_imp}
\usage{
nwos_estimates_geo_imp(
  geo.cd.list,
  imp,
  own = 45,
  domain,
  quest = QUEST,
  quest.meta = QUEST_META,
  plot = PLOTS,
  variables.categorcial = VARIABLES_CATEGORICAL,
  variables.continuous = VARIABLES_CONTINUOUS
)
}
\arguments{
\item{own}{= 45,}

\item{quest}{= QUEST,}

\item{quest.meta}{= QUEST_META,}

\item{plot}{= PLOTS,}

\item{variables.categorcial}{= VARIABLES_CATEGORICAL,}

\item{variables.continuous}{= VARIABLES_CONTINUOUS}
}
\value{
data frame
}
\description{
Generate estimates by geography and imputation
This is functuion is designed to be called by nwos_estimates_geo
}
\examples{
nwos_estimates_geo_imp()

}
