% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_geo_imp.R
\name{nwos_estimates_geo_imp}
\alias{nwos_estimates_geo_imp}
\title{nwos_estimates_geo_imp}
\usage{
nwos_estimates_geo_imp(
  geo.cd.list,
  imp,
  domain = "AC_WOOD >= 10",
  quest = QUEST,
  variables.categorcial = VARIABLES_CATEGORICAL,
  variables.continuous = VARIABLES_CONTINUOUS
)
}
\value{
data frame
}
\description{
Generate estimates by geography and imputation
This is functuion is designed to be called by nwos_estimates_geo
}
\examples{
nwos_estimates_geo_imp("1", 1)
nwos_estimates_geo_imp("1", 2)
mapply(nwos_estimates_geo_imp, 1, imp = 1:2, SIMPLIFY = F)

}
