% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_rep.R
\name{nwos_estimates_rep}
\alias{nwos_estimates_rep}
\title{NWOS Estimates Total by State, Variable, and Level}
\usage{
nwos_total_svl(state, variable, level, data = QUEST, area = NA, weights = "WEIGHT")
}
\arguments{
\item{state}{= "1"}

\item{variable}{= "HOME"}

\item{level}{= "1",}

\item{data}{= QUEST_WIDE}

\item{area}{= NA}

\item{weights}{= "WEIGHT"}
}
\description{
Estimator used to calculate NWOS totals by state, variable and level.
}
\examples{
rep = "0"
variable.list = VARIABLE_LIST
quest.list = QUEST_LIST
rep.list = REPLICATE_LIST
geo.list = GEO_LIST
stratum = "FFO"
domain = "TENPLUS"
i = 1
g = 1
v = 1
start.time <- Sys.time()
nwos_estimates_rep(geo.list = GEO_LIST[1,])
Sys.time() - start.time # Time difference of 11.77129 secs
#' start.time <- Sys.time()
nwos_estimates_rep(geo.list = GEO_LIST[1,])
Sys.time() - start.time # Time difference of 11.77129 secs
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
