% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_rep.R
\name{nwos_estimates_rep}
\alias{nwos_estimates_rep}
\title{NWOS Estimates Total by State, Variable, and Level}
\usage{
nwos_total_svl(state, variable, level, data = QUEST, area = NA, weights = "WEIGHT")
}
\arguments{
\item{state}{= "1"}

\item{variable}{= "HOME"}

\item{level}{= "1",}

\item{data}{= QUEST_WIDE}

\item{area}{= NA}

\item{weights}{= "WEIGHT"}
}
\description{
Estimator used to calculate NWOS totals by state, variable and level.
}
\examples{
rep = 1:2
variable.list = VARIABLE_LIST
quest.list = QUEST_LIST
rep.list = REPLICATES_LIST
geo.list = GEO_LIST
stratum = "FFO"
domain = "ONEPLUS"
g = 1
v = 1
x <- nwos_estimates_rep(variable.list = VARIABLE_LIST[1:5,], geo.list = GEO_LIST[1,])
nwos_estimates_rep(variable.list = VARIABLE_LIST[524,], geo.list = GEO_LIST[1,])
start.time <- Sys.time()
nwos_estimates_rep(geo.list = GEO_LIST[1,])
Sys.time() - start.time
start.time <- Sys.time()
x <- lapply(as.character(0:10), nwos_estimates_rep, variable.list = VARIABLE_LIST[1,], geo.list = GEO_LIST[1:2,])
Sys.time() - start.time
start.time <- Sys.time()
x <- mclapply(as.character(0:10), nwos_estimates_rep, variable.list = VARIABLE_LIST[1,], geo.list = GEO_LIST[1,], mc.cores = 3)
Sys.time() - start.time
bind_rows(mapply(variable = "TOTAL", level = 1, nwos_estimates_variable_imp_rep, SIMPLIFY = F))
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
