% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_summary.R
\name{nwos_estimates_summary}
\alias{nwos_estimates_summary}
\title{nwos_estimates_summary}
\usage{
nwos_estimates_summary(geo.abb, stratum = "FFO", domain = "TENPLUS")
}
\description{
Combine and ummarize Imputations
}
\examples{
domain = "FFO"
stratum = "ONEPLUS"
# Need to run nwos_estimates_combine first
geo.abb <- GEO_LIST$GEO_ABB
nwos_estimates_summary(geo.abb[55])
mclapply(geo.abb, nwos_estimates_summary, mc.cores = N_CORES)

}
