% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_summary AWS.R,
%   R/nwos_estimates_summary.R
\name{nwos_estimates_summary}
\alias{nwos_estimates_summary}
\title{nwos_estimates_summary}
\usage{
nwos_estimates_summary(geo.abb, stratum = "FFO", domain = "TENPLUS",
  wd = "DATA")

nwos_estimates_summary(geo.abb, stratum = "FFO", domain = "TENPLUS",
  wd = "DATA")
}
\description{
Combine and ummarize Imputations

Combine and ummarize Imputations
}
\examples{
library(tidyverse)
library(parallel)
list.files("/home/ubuntu/GEO")
domain = "FFO"
stratum = "ONEPLUS"
# Need to run nwos_estimates_combine first
GEO_LIST <- read.csv("GEO_LIST.csv", stringsAsFactors = F)
geo.abb <- GEO_LIST$GEO_ABB
nwos_estimates_summary(geo.abb[1])
N_CORES <- detectCores() - 1
mclapply(geo.abb, nwos_estimates_summary, mc.cores = N_CORES)

domain = "FFO"
stratum = "ONEPLUS"
# Need to run nwos_estimates_combine first
geo.abb <- GEO_LIST$GEO_ABB
nwos_estimates_summary(geo.abb[55])
mclapply(geo.abb, nwos_estimates_summary, mc.cores = N_CORES)

}
