% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_export.R
\name{nwos_export}
\alias{nwos_export}
\title{nwos_export}
\usage{
nwos_export(x, dir = "~", format = "wide")
}
\arguments{
\item{x}{is a nwos.object}

\item{dir}{is the directory to export to}

\item{format}{is a character string containing the values 'wide' or 'long'}
}
\value{
a data.frame
}
\description{
Exports NWOS data and metadata as CSVs, in either long or wide format.
}
\examples{
nwos_export(foo,dir='~',format='long')

}
