% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_full.R
\name{nwos_full}
\alias{nwos_full}
\title{nwos_full}
\usage{
nwos_full(x, imputations = "none")
}
\arguments{
\item{x}{is a nwos.object}

\item{imputations}{is string containing which imputation set is to be used in place of non-response, if any. Valid responses are "none","random","1","2","3","4","5"}
}
\value{
a data.frame
}
\description{
Formats full NWOS dataset (i.e. all plots and responses) into a wide format, for use as input in the 'survey' package
}
\examples{
nwos_full(foo,imputations="1")

}
