% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_full_metadata.R
\name{nwos_full_metadata}
\alias{nwos_full_metadata}
\title{nwos_full_metadata}
\usage{
nwos_full_metadata(quest, plots)
}
\arguments{
\item{quest}{is a nwos.object}

\item{plots}{is a nwos.plots.object}
}
\value{
a data.frame
}
\description{
Creates metadata table for full (i.e. responses and plots) NWOS dataset
}
\examples{
nwos_full_metadata(foo)

}
