% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_long.R
\name{nwos_long}
\alias{nwos_long}
\title{nwos_long}
\usage{
nwos_long(x, imputations = "none")
}
\arguments{
\item{x}{is a nwos.object}

\item{imputations}{is string containing which imputation set is to be used in place of non-response, if any. Valid responses are "none","random","1","2","3","4","5"}
}
\value{
a data.frame
}
\description{
Formats NWOS data into a long format for analysis
}
\examples{
nwos_long(foo,imputations="1")

}
