% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_replicates.R
\name{nwos_replicates}
\alias{nwos_replicates}
\title{NWOS Replicates}
\usage{
nwos_replicates(index, point.count, R = 5000)
}
\arguments{
\item{index}{vector of names used to identify rows/observations in the data.}

\item{point.count}{the number of sample points associated with each observation.}

\item{R}{number of replicates or bootstraps. Default is 2500.}
}
\value{
List containing indices for the observations that make up each of the replicates.
}
\description{
Generate replicates that can be used for NWOS bootstrapping variance estimation.
}
\details{
This function needs to be run by state.
}
\examples{
WI_REPLICATES <- nwos_replicates(index = row.names(wi), point.count = wi$POINT_COUNT, R = 100)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
