% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_response_rate.R
\name{nwos_response_rate}
\alias{nwos_response_rate}
\title{NWOS Response Rate}
\usage{
nwos_response_rate(point.count, response)
}
\arguments{
\item{df}{data frame containing, at a minimum, the following variables: STATE_CD, LAND_USE, OWN_CD, POINT_COUNT, and RESPONSE}

\item{state.list}{vector of state codes to be evaluated.}

\item{land.use.list}{vector of land uses to be evaluated. The defaualt value is 1 (forest).}

\item{own.cd.list}{vector of ownership codes to be evaluated. The default value if 45 (family).}
}
\value{
data frame of response rates by state/land use/ownership strata
}
\description{
This function calculates response rate for the NWOS.
}
\examples{
load("data/nwos_data_sample.RData")
data <- NWOS_DATA_SAMPLE[NWOS_DATA_SAMPLE$SAMPLE==1,]
nwos_response_rate(data$POINT_COUNT, data$RESPONSE)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
