% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_response_rate.R
\name{nwos_response_rate}
\alias{nwos_response_rate}
\title{NWOS Response Rate}
\usage{
nwos_response_rate(stratum, point.count, response)
}
\arguments{
\item{stratum}{vector indicating inclusion (1) and exclusion (0) in the stratum of interest.}

\item{point.count}{vector of the number of sample points associated with each observation.}

\item{response}{vector indicating response (1) and non-response (0).}
}
\value{
Response rate in the stratum.
}
\description{
This function calculates response rates for the NWOS.
}
\details{
This function needs to be run by state.
}
\examples{
wi <- wi \%>\% mutate(RESPONSE = if_else(RESPONSE_PROPENSITY >= 0.5, 1, 0))
WI_FFO_RR <- nwos_response_rate(stratum = wi$FFO, point.count = wi$POINT_COUNT, response = wi$RESPONSE)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
