% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table.R
\name{nwos_table}
\alias{nwos_table}
\title{nwos_table}
\usage{
nwos_table(
  tab.num,
  data = table.data,
  stratum.abb = STRATUM_ABB,
  stratum.name = STRATUM_NAME,
  domain.abb = DOMAIN_ABB,
  domain.name = DOMAIN_NAME,
  year = YEAR,
  year.range = YEAR_RANGE
)
}
\description{
nwos_table
}
\details{
For area and cooperation rate tables see ...
nwos_table()
nwos_table_area()
nwos_table_coop()
}
