% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table_coop.R
\name{nwos_table_coop}
\alias{nwos_table_coop}
\title{nwos_table}
\usage{
nwos_table_coop(geo.abb, data = COOP, ref.geo = REF_GEO,
  ref.tab = REF_TABLE, year = YEAR, year.range = YEAR_RANGE)
}
\description{
Create the body of an NWOS core, area, or cooperataion rate table
}
\details{
nwos_table_coop("US")
}
