% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_estimates_make_wide.R,
%   R/nwos_table_make_wide.R
\name{nwos_table_make_wide}
\alias{nwos_table_make_wide}
\title{nwos_table_make_wide}
\usage{
nwos_table_make_wide(data)

nwos_table_make_wide(data)
}
\description{
Convert long NWOS estimates to wide and add descriptors

Create the body of an NWOS table
}
\examples{
REF_GEO <- read_csv("DATA/REF/REF_GEO.csv")
REF_VARIABLE <- read_csv("DATA/REF/REF_VARIABLE.csv")
REF_TABLE <- read_csv("DATA/REF/REF_TABLE.csv")
REF_LABEL <- read_csv("DATA/REF/REF_LABEL.csv", col_types = list(col_character(), col_character(), col_character(), col_double()))
x <- nwos_table_make_wide(data)

x <- nwos_table_make_wide(data = readRDS("INPUTS/ESTIMATES/NWOS_2018_FFO_TENPLUS_US.RDS"))

}
