% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table_number.R
\name{nwos_table_number}
\alias{nwos_table_number}
\title{nwos_round}
\usage{
nwos_table_number(x, d = 0, less.one = T)
}
\description{
Converts units (e.g., to thousands) and rounds simialr to base except it adds another digit for values less than 1
}
\details{
For area and cooperation rate tables see ...
nwos_table_number(1100)
nwos_table_number(900)
nwos_table_number(0)
nwos_table_number(NA)
}
