% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table_row.R
\name{nwos_table_row}
\alias{nwos_table_row}
\title{nwos.table.row}
\usage{
nwos_table_row(i, data)
}
\description{
Create the body of an NWOS table
}
\details{
For area and cooperation rate tables see ...
nwos.table.row(data.row = data[1,])
}
