% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_table_set.R
\name{nwos_table_set}
\alias{nwos_table_set}
\title{nwos_table_set}
\usage{
nwos_table_set(
  geo.abb,
  ref.geo = GEO_LIST,
  ref.table = REF_TABLE,
  stratum = STRATUM,
  stratum.abb = STRATUM_ABB,
  stratum.name = STRATUM_NAME,
  domain = DOMAIN,
  domain.abb = DOMAIN_ABB,
  domain.name = DOMAIN_NAME,
  year = YEAR,
  year.range = YEAR_RANGE
)
}
\description{
Create the body of an NWOS table
}
\details{
For area and cooperation rate tables see ...
nwos_table_body()
nwos_table_body_area()
nwos_table_body_coop()
}
\examples{
nwos_table_set(geo.abb = "US")
geo.abb = "US"
data = ESTIMATES_WIDE
ref.geo = GEO_LIST
ref.table = REF_TABLE
stratum.name = "Family Forest Ownerships"
domain.name = "10\\\\texttt{+} acres"
year = "2018"
year.range = "2017-2018"

}
