% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_variance.R
\name{nwos_variance}
\alias{nwos_variance}
\title{NWOS Variance}
\usage{
nwosVar(state.area, point.count, response, owner.area, stratum, domain, variable = 1,
replicates, index, stat = "TOTAL", units = "OWNERSHIPS")
}
\arguments{
\item{stat}{statistic of interest. Permissable values are "TOTAL", "mean" or "proportion". Default is "total".}

\item{units}{units of analysis. Permissable values are "OWNERSHIPS" or "AREA". Default is "OWNERSHIPS".}
}
\description{
This function calculates variances for NWOS statistics using a bootstrapping approach.
}
\examples{
NEED TO ADD
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
