% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_weights.R
\name{nwos_weights}
\alias{nwos_weights}
\title{NWOS Weights}
\usage{
nwos_weights(stratum, point.count, response, owner.area, stratum.area, stratum.area.correction = stratum.area, response.rate)
}
\arguments{
\item{stratum}{vector indicating inclusion (1) and exclusion (0) in the stratum of interest.}

\item{point.count}{vector of the number of sample points associated with each observation.}

\item{response}{vector indicating response (1) and non-response (0).}

\item{stratum.area}{area (of forestland) in the stratum of interest.}

\item{stratum.area.correction}{area (of forestland) in the stratum of interest to which the total stratum.area will be forced. The default value is stratum.area (i.e., no correction).}

\item{response.rate}{response rate for the stratum of interest. Value should range from 0 ot 1.}

\item{owner.area}{vector of the area (of forestland) owned by each ownership.}
}
\value{
vector of weights
}
\description{
This function returns the calculated weights based on the NWOS sample design.
}
\examples{
wi$WEIGHT <- nwos_weights(stratum = wi$FFO, point.count = wi$POINT_COUNT,
response = wi$RESPONSE, owner.area = wi$AC_WOOD,
stratum.area = WI_FFO_AREA, response.rate = WI_FFO_RR)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
