% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_weights_trimmed.R
\name{nwos_weights_trimmed}
\alias{nwos_weights_trimmed}
\title{NWOS Trim Weights}
\usage{
nwos_weights_trimmed(weights, method = "iqr_1.5")
}
\arguments{
\item{method}{method to be used for trimming. The only method currently implemented is "iqr_1.5".}

\item{weight}{vector of weights per ownership.}
}
\description{
This function calculates trimmed weights for the NWOS.
}
\examples{
nwos_weights_trimmed(weights = wi$WEIGHT)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
