% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_wide.R
\name{nwos_wide}
\alias{nwos_wide}
\title{nwos_wide}
\usage{
nwos_wide(x, imputations = "none")
}
\arguments{
\item{x}{is a nwos.object}

\item{imputations}{is string containing which imputation set is to be used in place of non-response, if any. Valid responses are "none","1","2","3","4","5"}
}
\value{
a data.frame
}
\description{
Formats NWOS data into a wide format for analysis
}
\examples{
nwos_wide(foo,imputations='2')

}
