% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWOSload64.R
\name{NWOSload64}
\alias{NWOSload64}
\title{NWOSload64}
\usage{
NWOSload64(tables)
}
\arguments{
\item{tables}{is a character vector containing names of update tables currently in the working environment.}
}
\description{
Loads update tables to the NWOS database from a 64-bit install of R
}
\details{
This function must be run on a machine with an ODBC connection (32-bit) to the USFS FIA production database through a user with read permissions.
}
\examples{
NWOSload64(tables=c('SAMPLE_UPDATE','OWNER_UPDATE'))

}
