% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_proportion.R
\name{nwosProportion}
\alias{nwosProportion}
\title{NWOS Proportion}
\usage{
nwosProportion(weight, point.count, domain, y, area, units="ownerships")
}
\arguments{
\item{weight}{weight for each observation.}

\item{y}{variable of interest. 1 indicates inclusion.}

\item{domain}{variable indicating whether ownership is in the domain of interest.}

\item{area}{vector of areas of forest land owned by sampled ownerships. NAs are permissable.}

\item{units}{units of analysis. Permissable values are "ownerships" or "area". Default is "ownerships".}
}
\description{
This function calculates totals for the NWOS.
}
\examples{
load("data/nwos_data_sample.RData")
df <- nwos.data.sample[nwos.data.sample$SAMPLE==1,]
df$WEIGHT <- nwosWeights(point.count=df$POINT_COUNT, area=df$ACRES_FOREST, area.total=1000)
# Proportion of ownerships with Y_1=1
nwosProportion(weight=df$WEIGHT, y=df$Y_1)
# Proportion of acres with Y_1=1
nwosProportion(weight=df$WEIGHT, y=df$Y_1, area=df$ACRES_FOREST, units="area")
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
