% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_variance.R
\name{nwosVar}
\alias{nwosVar}
\title{NWOS Variance}
\usage{
nwosVar(point.count, domain, y=1, area, response=1,
stratum.area, units="ownerships", stat="total", R=100)
}
\arguments{
\item{data}{data frame of survey data. See details below for required variables.}

\item{units}{units of analysis. Permissable values are "ownerships" or "area". Default is "ownerships".}

\item{stat}{statistic of interest. Permissable values are "total", "mean" or "proportion". Default is "total".}

\item{R}{number of replicates or bootstraps. Default is 100.}
}
\description{
This function calculates variances for NWOS statistics using a bootstrapping approach.
}
\examples{
load("data/nwos_data_sample.RData")
df <- nwos.data.sample[nwos.data.sample$SAMPLE==1,]
nwosVar(point.count=df$POINT_COUNT, area=df$ACRES_FOREST, area.total=1000, R=100)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
