% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwos_weights.R
\name{nwosWeights}
\alias{nwosWeights}
\title{NWOS Weights}
\usage{
nwosWeights(point.count, area, domain, stratum.area, response.rate)
}
\arguments{
\item{point.count}{vector of number of sample points. Needs to sum to total number of sample points across all land uses and ownership classes.}

\item{area}{vector of areas of forest land owned by sampled ownerships. NAs are permissable for ownerships not in owner.class.}

\item{stratum.area}{total area of land.}

\item{response.rate}{vector of response rates.}

\item{owner.class}{vector of owner classes.}

\item{response.rate.owner.class}{vector of owner.class associated with each response.rate}
}
\value{
vector of weights
}
\description{
This function returns the calculated weights based on the NWOS sample design.
}
\examples{
# Calculate weights for family forest ownerships
load("data/nwos_sample_data.RData")
nwos.sample.data$domain <- ifelse(nwos.sample.data$owner.class=="FamilyForest", 1, 0)
nwos.sample.data$response <- ifelse(nwos.sample.data$owner.class=="FamilyForest",
as.numeric(as.character(nwos.sample.data$response)), NA)
sample.response.rate <- nwosResponseRates(nwos.sample.data$point.count,nwos.sample.data$response)
nwos.sample.data$weights <- nwosWeights(point.count=nwos.sample.data$point.count,
area=nwos.sample.data$area,
domain=nwos.sample.data$domain,
stratum.area=35198019,
response.rate=sample.response.rate)
}
\references{
Butler, B.J. In review. Weighting for the US Forest Service, National Woodland Owner Survey. U.S. Department of Agriculture, Forest Service, Northern Research Station. Newotwn Square, PA.
}
\keyword{nwos}
